[rename] : rename test packge
