+++ src/main/java/kr/where/backend/group/dto/GroupMemberRequestDTO.java
@@ -1,16 +1,16 @@
 package kr.where.backend.group.dto;
 
-import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor
-public class GroupCreateRequestDTO {
+public class GroupMemberRequestDTO {
+    @NotNull
+    private Long groupId;
     @NotNull
     private Long memberId;
-
-    @NotEmpty
-    private String groupName;
+    @NotNull
+    private Boolean is_owner;
 }
