+++ src/main/java/kr/where/backend/group/entity/GroupMember.java
@@ -9,6 +9,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import kr.where.backend.member.Member;
+
 import lombok.Getter;
 import lombok.Setter;
